requirements.txt
pandas
boto3
Pillow
pydantic
streamlit-image-select
aiohttp
asyncio
python-slugify